==> notebooks/04_transforming_model_data.ipynb <==
# Scaling model data

Before moving on to model improvements, we will first scale the numerical data used for the model in the previous notebook. This is a necessary step for some models, and it can also improve the performance of others. We will use the `StandardScaler` from `sklearn` to scale the data.

Aditionally, we will also apply logarithmic transformation to a few variables. This is a common practice when dealing with skewed data, and it can also improve the performance of some models.

import pickle
import pathlib

import numpy as np
import pandas as pd

DATA_DIR = pathlib.Path.cwd().parent / 'data'
print(DATA_DIR)

model_data_path = DATA_DIR / 'processed' / 'ames_model_data.pkl'

with open(model_data_path, 'rb') as file:
    data = pickle.load(file)

data.info()

### Let´s import StandardScaler and create an instance of it:

from sklearn.preprocessing import StandardScaler

scaler = StandardScaler()

The StandardScaler scales each feature to have a mean of 0 and a standard deviation of 1. This is done by subtracting the mean and dividing by the standard deviation for each feature. The scaling is done independently for each feature, so that the features are on the same scale. This is important for some models, such as linear models, but not for others, such as decision trees.

We want to use the StandardScaler on continuous variables only, so we will first adapt the list of the continuous variables from previous notebooks:

continuous_variables = [
    'Lot.Frontage',
    'Lot.Area',
    'Mas.Vnr.Area',
    'BsmtFin.SF.1',
    'BsmtFin.SF.2',
    'Bsmt.Unf.SF',
    'Total.Bsmt.SF',
    'X1st.Flr.SF',
    'X2nd.Flr.SF',
    'Low.Qual.Fin.SF',
    'Gr.Liv.Area',
    'Garage.Area',
    'Wood.Deck.SF',
    'Open.Porch.SF',
    'Enclosed.Porch',
    'X3Ssn.Porch',
    'Screen.Porch',
    'Pool.Area',
    'Misc.Val',
]

data_cont = data.loc[:, continuous_variables]

data_cont.describe()

scaler.fit(data_cont)

scaled_data_cont = scaler.transform(data_cont)

### Create a new DataFrame with the scaled data:

scaled_data_cont = pd.DataFrame(scaled_data_cont, columns=data_cont.columns, index=data_cont.index)

scaled_data_cont.describe()

### Finally, join the scaled data with the categorical variables again:

data.loc[:, continuous_variables] = scaled_data_cont

data.describe()

### Now for the logarithmic transformation.

First we have to check the skewness of the numerical variables, in order to select the ones that need to be transformed. To do that, we will use the `skew()` method from `scipy.stats`. The skewness of a distribution is a measure of its asymmetry around the mean (a distribution is symmetric if it looks the same to the left and right of the center point). It is calculated as:

$$\frac{\sum_{i=1}^{n}(x_i - \bar{x})^3}{(n-1)\sigma^3}$$

Where $x_i$ is the "i"th observation, $\bar{x}$ is the mean, $\sigma$ is the standard deviation and $n$ is the number of observations.

from scipy.stats import skew

skewness = data.select_dtypes(include=[np.number]).apply(lambda x: skew(x))

skewness

The columns that have a skewness value greater than 5 will be selected for transformation

skewed_cols = skewness[skewness > 5].index

print(f"{len(skewed_cols)} out of {len(data.columns)} columns are skewed.")
print(skewed_cols)

We will create a logarithmic transformer using `FunctionTransformer` from `sklearn.preprocessing`. This transformer will apply the logarithmic transformation to the selected variables, adding 1 to each value before applying the logarithm, in order to avoid errors when the variable contains 0.

The values are clipped in the range [0, 1e10] to avoid negative values and overflow, which could result in a NaN value.

from sklearn.preprocessing import FunctionTransformer

def log_transform(X):
    return np.log1p(X.clip(lower=0, upper=1e10))

log_transformer = FunctionTransformer(func=log_transform, validate=False)

data.loc[:, skewed_cols] = log_transformer.transform(data.loc[:, skewed_cols])

data.describe()

Confirming that there are still no missing values:

data.isnull().sum()[data.isnull().sum() > 0]

### Saving the data for the next notebook:

model_data_scaled_path = DATA_DIR / 'processed' / 'ames_model_data_scaled.pkl'

with open(model_data_scaled_path, 'wb') as file:
    pickle.dump(data, file)
